BUG: Update pandas-gbq to latest version of google-cloud-bigquery
There was a breaking change in 0.32.0.dev1 which changed the way
configuration for the query job gets loaded. Also, it added the
'description' field to the schema resource, so this change updates the
schema comparison logic to account for that.

Updates the MASTER build in CI to also build with google-cloud-bigquery
at MASTER.