CLN: rename internal modules. move tests. (#163)
* CLN: rename internal modules. move tests.

Per GH#154 we don't need to use underscore to show submodules are
private. To be 100% clear, I have added a disclaimer to the API
reference as well.

Also, I have split the system tests from the unit tests so that the
(fast) unit tests can be more easily run separately from the (slow)
system tests. I have followed the same directory structure as used in
the google-cloud-bigquery library.

* CLN: move _version.py back for versioneer

* CLN: run tests from tests directory in conda test run