CLN: use pydata-google-auth for auth flow
Only private_key logic and customized path for credentials cache remain.
At some point in the future private_key logic will be removed, as that
parameter is deprecated in favor of the credentials argument.

Also removes the _try_credentials logic, as that slows down the
authentication process and is largely unnecessary now that credentials
can be explicitly created and supplied via the credentials argument.