read project_id from credentials if available (#258)
* read project_id from credentials if available

* Move project inference from credentials to shared GbqConnector

To more easily test the project ID inference, this commit also sends
the project ID as part of the query method call. Even though this is
redundant because the client object also contains the project ID, we
aren't exercising the bigquery.Client constructor in the unit tests to
avoid making real API calls.