BUG: Ensure table_schema arg not modified inplace
Fixes #277.

If any dictionary entry in the `table_schema` arg did
not contain a "mode" key, a mode="NULLABLE" kv pair
would be created; because `schema.update_schema()`
returns an object with references to its mutable input,
this allows the argument to ultimately be modified
by the function rather than the caller.

This pattern was used in both gbq.py and load.py,
so it was refactored into a helper function in
schema.py, which now returns a modified *copy*.
Deepcopy is used because the input is a list of
dictionaries, so a shallow copy would be insufficient.